+++ debug/svp.js
@@ -32,8 +32,200 @@ var io_error_default = IOError;
 
 // build/class/io_elements.js
 var IOElements = function() {
-  function IOElements2(options, identifiersClass, identifiersId, tagVideo) {
-    this.iconsPath = {
+  function IOElements2(options, identifiersClass, identifiersId, identifiersActions, identifiersIcons, elementVideo) {
+    this.elementContainter = document.createElement("div");
+    this.options = options;
+    this.identifiersClass = identifiersClass;
+    this.identifiersId = identifiersId;
+    this.identifiersActions = identifiersActions;
+    this.identifiersIcons = identifiersIcons;
+    this.elementVideo = elementVideo;
+    this.build();
+  }
+  IOElements2.prototype.build = function() {
+    this.elementContainter.setAttribute("id", this.identifiersId.container);
+    if (this.elementVideo.parentElement == null) {
+      new io_error_default("this.elementVideo.parentElement is null");
+    }
+    this.elementVideo.parentElement.insertBefore(this.elementContainter, this.elementVideo);
+    this.elementContainter.appendChild(this.elementVideo);
+    this.buildTop();
+    this.buildMiddle();
+    this.buildBottom();
+  };
+  IOElements2.prototype.buildTop = function() {
+    var elementTop = document.createElement("div");
+    elementTop.setAttribute("id", this.identifiersId.top);
+    this.elementContainter.appendChild(elementTop);
+    if (true) {
+      elementTop.innerHTML = "";
+    }
+  };
+  IOElements2.prototype.buildMiddle = function() {
+    var elementMiddle = document.createElement("button");
+    elementMiddle.setAttribute("id", this.identifiersId.middle);
+    elementMiddle.setAttribute("class", this.identifiersClass.buttons);
+    elementMiddle.setAttribute("action", this.identifiersActions.playPause);
+    this.elementContainter.appendChild(elementMiddle);
+    elementMiddle.innerHTML = this.buildIcon(this.identifiersIcons.play);
+  };
+  IOElements2.prototype.buildBottom = function() {
+    var elementBottom = document.createElement("div");
+    elementBottom.setAttribute("id", this.identifiersId.bottom);
+    this.elementContainter.appendChild(elementBottom);
+    var buttonPlayPause = document.createElement("button");
+    buttonPlayPause.setAttribute("class", this.identifiersClass.buttons);
+    buttonPlayPause.setAttribute("action", this.identifiersActions.playPause);
+    buttonPlayPause.innerHTML = this.buildIcon(this.identifiersIcons.play);
+    elementBottom.appendChild(buttonPlayPause);
+    var pCurrent = document.createElement("p");
+    pCurrent.setAttribute("action", this.identifiersActions.currentTime);
+    elementBottom.appendChild(pCurrent);
+    var divRangerProguessContainer = document.createElement("div");
+    divRangerProguessContainer.setAttribute("action", this.identifiersActions.rangerProguessContainer);
+    elementBottom.appendChild(divRangerProguessContainer);
+    var divRangerProguessInput = document.createElement("input");
+    divRangerProguessInput.setAttribute("action", this.identifiersActions.rangerProguessInput);
+    divRangerProguessInput.setAttribute("type", "range");
+    divRangerProguessInput.setAttribute("value", "0");
+    divRangerProguessInput.setAttribute("min", "0");
+    divRangerProguessInput.setAttribute("max", "100");
+    divRangerProguessContainer.appendChild(divRangerProguessInput);
+    var divRangerProguessDiv = document.createElement("div");
+    divRangerProguessDiv.setAttribute("action", this.identifiersActions.rangerProguessDiv);
+    divRangerProguessContainer.appendChild(divRangerProguessDiv);
+    var pDuration = document.createElement("p");
+    pDuration.setAttribute("action", this.identifiersActions.durationTime);
+    elementBottom.appendChild(pDuration);
+    var buttonVolume = document.createElement("button");
+    buttonVolume.setAttribute("class", this.identifiersClass.buttons);
+    buttonVolume.setAttribute("action", this.identifiersActions.volume);
+    buttonVolume.innerHTML = this.buildIcon(this.identifiersIcons.volumeOn);
+    elementBottom.appendChild(buttonVolume);
+    var buttonPictureInPicture = document.createElement("button");
+    buttonPictureInPicture.setAttribute("class", this.identifiersClass.buttons);
+    buttonPictureInPicture.setAttribute("action", this.identifiersActions.pictureInPicture);
+    buttonPictureInPicture.innerHTML = this.buildIcon(this.identifiersIcons.pictureInPicture);
+    elementBottom.appendChild(buttonPictureInPicture);
+    var buttonFullscreen = document.createElement("button");
+    buttonFullscreen.setAttribute("class", this.identifiersClass.buttons);
+    buttonFullscreen.setAttribute("action", this.identifiersActions.fullscreen);
+    buttonFullscreen.innerHTML = this.buildIcon(this.identifiersIcons.fullscreenOff);
+    elementBottom.appendChild(buttonFullscreen);
+  };
+  IOElements2.prototype.buildIcon = function(pathIcon) {
+    return '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 -960 960 960" class="'.concat(this.identifiersClass.icons, '"><path d="').concat(pathIcon, '" /></svg>');
+  };
+  return IOElements2;
+}();
+var io_elements_default = IOElements;
+
+// build/class/io_controllers.js
+var IOControllers = function() {
+  function IOControllers2(options, identifiersClass, identifiersId, identifiersActions, identifiersIcons, elementVideo) {
+    this.controlsValue = {
+      current: "00:00",
+      currentTime: 0,
+      duration: "00:00",
+      durationTime: 0
+    };
+    this.options = options;
+    this.identifiersClass = identifiersClass;
+    this.identifiersId = identifiersId;
+    this.identifiersActions = identifiersActions;
+    this.identifiersIcons = identifiersIcons;
+    this.elementVideo = elementVideo;
+    this.controlsElements = {
+      currentTime: document.querySelector('p[action="'.concat(this.identifiersActions.currentTime, '"]')),
+      durationTime: document.querySelector('p[action="'.concat(this.identifiersActions.durationTime, '"]'))
+    };
+    this.build();
+  }
+  IOControllers2.prototype.build = function() {
+    this.buildPrepareVideo();
+    this.buildPrepareVideoTime();
+  };
+  IOControllers2.prototype.buildPrepareVideo = function() {
+    this.elementVideo.removeAttribute("controls");
+    this.elementVideo.removeAttribute("height");
+    this.elementVideo.removeAttribute("width");
+    this.elementVideo.controls = false;
+    this.elementVideo.currentTime = this.options.currentTime;
+    this.elementVideo.setAttribute("playsinline", "");
+    this.elementVideo.setAttribute("preload", "auto");
+    this.elementVideo.setAttribute("controlslist", "nodownload nofullscreen noremoteplayback");
+    this.elementVideo.setAttribute("disablepictureinpicture", "");
+    this.elementVideo.setAttribute("disableremoteplayback", "");
+    this.elementVideo.setAttribute("disabledownload", "");
+    this.elementVideo.addEventListener("contextmenu", function(event) {
+      event.preventDefault();
+    });
+    if (this.options.autoplay) {
+      this.elementVideo.setAttribute("autoplay", "");
+    } else {
+      this.elementVideo.removeAttribute("autoplay");
+    }
+  };
+  IOControllers2.prototype.buildPrepareVideoTime = function() {
+    var _this = this;
+    this.elementVideo.onloadedmetadata = function() {
+      var _a;
+      _this.controlsElements.currentTime.innerHTML = _this.controlsValue.current;
+      _this.controlsValue.durationTime = _this.elementVideo.duration;
+      _this.controlsValue.duration = _this.formatTime((_a = _this.elementVideo.duration) !== null && _a !== void 0 ? _a : 0);
+      _this.controlsElements.durationTime.innerHTML = _this.controlsValue.duration;
+    };
+  };
+  IOControllers2.prototype.formatTime = function(seconds) {
+    function padZero(number) {
+      return (number < 10 ? "0" : "") + number;
+    }
+    var hours = Math.floor(seconds / 3600);
+    var minutes = Math.floor(seconds % 3600 / 60);
+    var secondsLeft = Math.floor(seconds % 60);
+    if (hours > 0) {
+      return padZero(hours) + ":" + padZero(minutes) + ":" + padZero(secondsLeft);
+    }
+    return padZero(minutes) + ":" + padZero(secondsLeft);
+  };
+  return IOControllers2;
+}();
+var io_controllers_default = IOControllers;
+
+// build/svp.js
+var SVP = function() {
+  function SVP2(options) {
+    var _a, _b, _c;
+    this.optionsDefault = {
+      autoplay: false,
+      volume: 70,
+      currentTime: 0
+    };
+    this.identifiersClass = {
+      all: "svp__all",
+      buttons: "svp__buttons",
+      icons: "svp__icons",
+      fading: "svp__fading",
+      hide: "svp__hide"
+    };
+    this.identifiersId = {
+      container: "svp__container",
+      top: "svp__top",
+      middle: "svp__middle",
+      bottom: "svp__bottom"
+    };
+    this.identifiersActions = {
+      playPause: "svp__action__playPause",
+      fullscreen: "svp__action__fullscreen",
+      volume: "svp__action__volume",
+      durationTime: "svp__action__durationTime",
+      currentTime: "svp__action__currentTime",
+      pictureInPicture: "svp__action__pictureInPicture",
+      rangerProguessContainer: "wvp__action__rangerProguessContainer",
+      rangerProguessInput: "wvp__action__rangerProguessInput",
+      rangerProguessDiv: "wvp__action__rangerProguessDiv"
+    };
+    this.indentifersIcons = {
       fullscreenOn: "M333-200v-133H200v-60h193v193h-60Zm234 0v-193h193v60H627v133h-60ZM200-567v-60h133v-133h60v193H200Zm367 0v-193h60v133h133v60H567Z",
       fullscreenOff: "M200-200v-193h60v133h133v60H200Zm0-367v-193h193v60H260v133h-60Zm367 367v-60h133v-133h60v193H567Zm133-367v-133H567v-60h193v193h-60Z",
       pause: "M560-200v-560h160v560H560Zm-320 0v-560h160v560H240Z",
@@ -42,72 +234,26 @@ var IOElements = function() {
       volumeOn: "M560-131v-62q97-28 158.5-107.5T780-481q0-101-61-181T560-769v-62q124 28 202 125.5T840-481q0 127-78 224.5T560-131ZM120-360v-240h160l200-200v640L280-360H120Zm420 48v-337q55 17 87.5 64T660-480q0 57-33 104t-87 64ZM420-648 307-540H180v120h127l113 109v-337Zm-94 168Z",
       volumeOff: "M813-56 681-188q-28 20-60.5 34.5T553-131v-62q23-7 44.5-15.5T638-231L473-397v237L273-360H113v-240h156L49-820l43-43 764 763-43 44Zm-36-232-43-43q20-34 29.5-71.923T773-481q0-103.322-60-184.661T553-769v-62q124 28 202 125.5T833-481q0 51-14 100t-42 93ZM643-422l-90-90v-130q47 22 73.5 66t26.5 96q0 15-2.5 29.5T643-422ZM473-592 369-696l104-104v208Zm-60 286v-150l-84-84H173v120h126l114 114Zm-42-192Z"
     };
-    this.options = options;
-    this.identifiersClass = identifiersClass;
-    this.identifiersId = identifiersId;
-    this.tagVideo = tagVideo;
-    this.tagContainter = document.createElement("div");
-    this.build();
-  }
-  IOElements2.prototype.build = function() {
-    this.tagContainter.setAttribute("id", this.identifiersId.container);
-    if (this.tagVideo.parentElement == null) {
-      new io_error_default("this.tagVideo.parentElement is null");
-    }
-    this.tagVideo.parentElement.insertBefore(this.tagContainter, this.tagVideo);
-    this.tagContainter.appendChild(this.tagVideo);
-  };
-  IOElements2.prototype.name = function() {
-  };
-  return IOElements2;
-}();
-var io_elements_default = IOElements;
-
-// build/svp.js
-var SVP = function() {
-  function SVP2(options) {
-    var _a, _b, _c, _d;
-    this.optionsDefault = {
-      settings: {
-        autoplay: false
-      },
-      video: {
-        volume: 70
-      }
-    };
-    this.identifiersClass = {
-      all: "svp_all",
-      buttons: "svp__buttons",
-      icons: "svp__icons",
-      fading: "svp__fading",
-      hide: "svp__hide"
-    };
-    this.identifiersId = {
-      container: "svp__container",
-      video: "svp",
-      top: "svp__top",
-      middle: "svp__middle",
-      bottom: "svp__bottom"
-    };
     this.options = {
-      settings: {
-        autoplay: (_b = (_a = options === null || options === void 0 ? void 0 : options.settings) === null || _a === void 0 ? void 0 : _a.autoplay) !== null && _b !== void 0 ? _b : this.optionsDefault.settings.autoplay
-      },
-      video: {
-        volume: (_d = (_c = options === null || options === void 0 ? void 0 : options.video) === null || _c === void 0 ? void 0 : _c.volume) !== null && _d !== void 0 ? _d : this.optionsDefault.video.volume
-      }
+      autoplay: (_a = options === null || options === void 0 ? void 0 : options.autoplay) !== null && _a !== void 0 ? _a : this.optionsDefault.autoplay,
+      volume: (_b = options === null || options === void 0 ? void 0 : options.volume) !== null && _b !== void 0 ? _b : this.optionsDefault.volume,
+      currentTime: (_c = options === null || options === void 0 ? void 0 : options.currentTime) !== null && _c !== void 0 ? _c : this.optionsDefault.currentTime
     };
     if (this.options == void 0) {
       new io_error_default("Options undefinded");
     }
-    this.tagVideo = document.getElementById(this.identifiersId.video);
-    if (this.tagVideo == void 0) {
+    this.tagVideos = document.querySelectorAll('video[plugin="svp"]');
+    if (this.tagVideos == void 0) {
       new io_error_default("Tag Video undefinded");
     }
     this.build();
   }
   SVP2.prototype.build = function() {
-    new io_elements_default(this.options, this.identifiersClass, this.identifiersId, this.tagVideo);
+    var _this = this;
+    this.tagVideos.forEach(function(elementVideo) {
+      new io_elements_default(_this.options, _this.identifiersClass, _this.identifiersId, _this.identifiersActions, _this.indentifersIcons, elementVideo);
+      new io_controllers_default(_this.options, _this.identifiersClass, _this.identifiersId, _this.identifiersActions, _this.indentifersIcons, elementVideo);
+    });
   };
   return SVP2;
 }();