+++ src/interfaces/indentifers_actions.ts
@@ -1,0 +1,12 @@
+// ==================================================
+export default interface identifiersActions {
+    playPause: string,
+    fullscreen: string,
+    volume: string,
+    durationTime: string,
+    currentTime: string,
+    pictureInPicture: string,
+    rangerProguessContainer: string,
+    rangerProguessInput: string,
+    rangerProguessDiv: string,
+}