+++ src/interfaces/indentifers_icons.ts
@@ -1,0 +1,10 @@
+// ==================================================
+export default interface IndentifersIcons {
+    fullscreenOn: string,
+    fullscreenOff: string,
+    pause: string,
+    play: string,
+    pictureInPicture: string,
+    volumeOn: string,
+    volumeOff: string,
+}