+++ src/interfaces/indentifers_id.ts
@@ -1,7 +1,6 @@
 // ==================================================
 export default interface identifiersId {
     container: string,
-    video: string,
     top: string,
     middle: string,
     bottom: string,