+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # 1.0.0
 
     - First version(Experimental)
+
+# 1.0.1 
+    - Add new scripts build
+    - Addons debug
+    - Bug fixes